requirements.txt
pymupdf
fastembed
faiss-cpu
pydantic
langchain
langchain-community
fastapi
uvicorn
streamlit
typer
loguru
python-dateutil
langdetect
gradio
